Automatic Object Recognition and Description.md
The need for automated systems has become very important and rampant due to the current trends in technology. There exists already trained models capable of predicting global images, but not so much of those found in this part of the globe. There then arises the need to develop a model that can predict with a reasonable degree of accuracy objects in this vicinity as detection models can now be integrated into surveillance systems generally improving the security of the inhabitants of this area. A little research has been done in this area, but they did not incorporate very local images like the one’s found in almost every Nigerian kitchen. The local image detection software is a software that was developed to classify 12 different types of images found in this locality. They include: Apples, Oranges, Brooms, Onga stew seasoning and 8 other classes. One of the major aims of the project is to add to the research community (a dataset containing local images and a model that can predict these images). The convolutional neural architecture VGG16 was used in a technique called transfer learning to use weights from ImageNet as the initial layers of our model. It returned a reasonable accuracy and so we didn’t bother training on another architecture. If this software is upgraded to incorporate life detection, it can be incorporated into a home surveillance system, and can aid a complete life surveillance of an area, making it easy to find things at the exact point they are located if they are within the line of sight of the surveillance system.